Move boardslice
