+++ src/redux/column.slice.ts
@@ -2,7 +2,7 @@ import { createSlice } from "@reduxjs/toolkit";
 import { Column as innerColumn } from "src/types/column.g";
 import { Column as OuterColumn } from "src/features/board/board.api";
 import { RootState } from "src/redux/root";
-import { getBoardThunk } from "src/redux/board.slice";
+import { getBoardThunk } from "src/features/board/board.slice";
 
 interface ColumnState {
   columns: innerColumn[];
