+++ src/redux/root.ts
@@ -1,7 +1,7 @@
 import { combineReducers } from "redux";
 import { authReducer } from "src/features/auth/auth.slice";
 import { userReducer } from "src/redux/user.slice";
-import { boardReducer } from "src/redux/board.slice";
+import { boardReducer } from "src/features/board/board.slice";
 import { columnReducer } from "src/redux/column.slice";
 import { cardReducer } from "src/redux/card.slice";
